requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic>=2.10,<3
pydantic-settings>=2.6.1
httpx==0.27.2
python-dotenv==1.0.1

# Azure SDKs
azure-identity==1.19.0
azure-mgmt-subscription==3.1.1
azure-mgmt-monitor==6.0.2
azure-mgmt-costmanagement==4.0.1
msgraph-sdk==1.12.0

# LLM / Agent
langchain==0.3.7
langchain-openai==0.2.8
langgraph==0.2.45
langchain-mcp-adapters==0.0.3

# MCP
mcp>=1.2.1

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
ruff>=0.9.3
